Fix int64 cast for writeUnicodeCodepoint call in \u handler
The \U handler already receives int64 from parseHex, but the \u handler
also returns int64 and needs an explicit cast for writeUnicodeCodepoint.

+++ interp/builtins/echo/echo.go
@@ -184,7 +184,7 @@ func processEscapes(s string) (string, bool) {
 				continue
 			}
 			i += consumed
-			writeUnicodeCodepoint(&b, val)
+			writeUnicodeCodepoint(&b, int64(val))
 			continue
 		case 'U':
 			// Unicode: \UHHHHHHHH (up to 8 hex digits)
